refactor: move operations table layout to shared
