+++ src/renderer/shared/ui/operations-table-layout.ts
